+++ src/forms/index.js
@@ -2,9 +2,12 @@
 
 // Define a submodule
 var angular = require('angular');
-angular.module('md.forms', []);
 
+var dependencies = [
+  'ui.router',
+  'restangular'
+];
+angular.module('md.forms', dependencies);
 
 require('./directives');
 require('./services');
-